+++ app/src/main/java/com/example/blockcanary/DemoFragment.java
@@ -1,0 +1,105 @@
+package com.example.blockcanary;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+
+
+public class DemoFragment extends Fragment implements View.OnClickListener{
+
+    public static DemoFragment newInstance()
+    {
+        DemoFragment f = new DemoFragment();
+        return  f;
+    }
+
+    @Override
+    public void onCreate(final Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(final LayoutInflater inflater, final ViewGroup container, final Bundle savedInstanceState) {
+        View  v = inflater.inflate(R.layout.activity_main,null);
+        return  v;
+    }
+
+    @Override
+    public void onViewCreated(final View view, @Nullable final Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        Button button1 = (Button) view.findViewById(R.id.button1);
+        Button button2 = (Button) view.findViewById(R.id.button2);
+        Button button3 = (Button) view.findViewById(R.id.button3);
+
+        button1.setOnClickListener(this);
+        button2.setOnClickListener(this);
+        button3.setOnClickListener(this);
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable final Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+    }
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.button1:
+                try {
+                    Thread.sleep(2000);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+                break;
+            case R.id.button2:
+                for (int i = 0; i < 100; ++i) {
+                    readFile();
+                }
+                break;
+            case R.id.button3:
+                double result = compute();
+                System.out.println(result);
+                break;
+        }
+    }
+
+    private double compute() {
+        double result = 0;
+        for (int i = 0; i < 1000000; ++i) {
+            result += Math.acos(Math.cos(i));
+            result -= Math.asin(Math.sin(i));
+        }
+        return result;
+    }
+
+    private void readFile() {
+        FileInputStream reader = null;
+        try {
+            reader = new FileInputStream("/proc/stat");
+            while (reader.read() != -1);
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (reader != null) {
+                try {
+                    reader.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+    }
+}
