+++ app/src/main/res/layout/activity_demo.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:tools="http://schemas.android.com/tools"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:fitsSystemWindows="false"
+              android:orientation="vertical">
+
+    <View
+        android:id="@+id/view_state_bar"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/state_bar_height"
+        android:background="@color/material_blue" />
+
+    <FrameLayout
+        android:layout_below="@+id/view_state_bar"
+        android:id="@+id/container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:ignore="MergeRootFrame" />
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentRight="true"
+        android:layout_margin="16dp"
+        android:src="@drawable/ic_done"/>
+</RelativeLayout>
