+++ app/src/main/res/values-v19/styles.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <style name="CjjBaseTheme" parent="Base.Theme.DesignDemo">
+        <item name="android:windowTranslucentStatus">true</item>
+    </style>
+
+</resources>
