+++ app/src/main/res/values/styles.xml
@@ -3,7 +3,7 @@
     <style name="CjjBaseTheme" parent="Base.Theme.DesignDemo">
     </style>
 
-    <style name="Base.Theme.DesignDemo" parent="Theme.AppCompat.Light.DarkActionBar">
+    <style name="Base.Theme.DesignDemo" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="colorPrimary">#2196F3</item>
         <item name="colorPrimaryDark">#1976D2</item>
         <item name="colorAccent">#F44336</item>
